[1/2] SystemUI: Add time-context headers to notification header
This commit adds image headers to notification panel header, based on the date/time.

There is a default provider that provides sky pictures as well as special events like Christmas and new year's eve.

More will be coming, with eventually a way to use external (as in, Internet) provider for those pictures.

Change-Id: I0d43bc3fa74880d1a09b2eebe2c27aa419c43a5c
Signed-off-by: W4TCH0UT <ateekujjawal@gmail.com>

Conflicts:
	res/values/tr_strings.xml
	res/xml/notification_drawer_settings.xml